+++ Bot/Database/Types/Public/UsersRow.cs
@@ -115,7 +115,7 @@ public class UsersRow(string connectionString, HandlersGroup handlersGroup, IDat
         await using NpgsqlConnection connection = await GetConnectionAsync();
         await using NpgsqlCommand command = connection.CreateCommand();
         command.CommandText = "DELETE FROM public.users WHERE id = @id;";
-        command.Parameters.Add(new NpgsqlParameter("id", DbType.VarNumeric) { Value = Id });
+        command.Parameters.Add(new NpgsqlParameter("id", DbType.VarNumeric) { Value = (long)Id });
 
         await command.ExecuteNonQueryAsync();
     }
